chnages

+++ src/payment/models.py
@@ -5,7 +5,7 @@ from django.core.exceptions import ValidationError
 
 class Payment(models.Model):
     account = models.ForeignKey(Account, on_delete=models.CASCADE)
-    amount = models.DecimalField(max_digits=10, decimal_places=2)
+    amount = models.IntegerField()
     upi_id = models.CharField(max_length=100)
     created_at = models.DateTimeField(auto_now_add=True)
     status = models.CharField(max_length=20, choices=[
